+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `Log`: Remove unwanted format tags from non-colored output
 -   `App::invokeLogging()`: Fix default log level from invalid value `-1` to `31` (everything but debug)
+-   `App::parseArguments()`: Do not miss arguments in combination with options
 -   `Command\Command::tearDown()`: Print last debug message (spent time, consumed memory etc.) to STDERR
 -   `composer`: Clarify which PHP extensions are required and which ones are suggested